step1/src/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Shakesapp server"
authors = ["Yoshi Yamaguchi <yoshifumi@google.com>"]
license = "Apache-2.0"
include = [
  "server.py",
  "shakesapp_pb2.py",
  "shakesapp_pb2_grpc.py",
]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.36.1"
protobuf = "^3.15.5"
grpcio-health-checking = "^1.36.1"
google-cloud-storage = "^1.36.1"
structlog = "^21.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.7.0"
grpcio-tools = "^1.36.1"

[tool.black]
extend-exclude = '''
^/(
    shakesapp_pb2_grpc.py
  | shakesapp_pb2.py
)
'''

[tool.poetry.scripts]
serve = 'server:serve'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
